Use the NNAPI memory to receive the remote inference output
The output tensor from OpenVino was being used to receive output
data from the DetectionClient. This was being done to allow any
required internal data conversions.

Since AI Dispatcher is now able to send the output data in the
desired format, the pointer to NNAPI memory can be used directly
to accept it.

Tracked-On: OAM-106985
Signed-off-by: Anoob Anto K <anoob.anto.kodankandath@intel.com>